Moved core logic into its own file - CloudFlare/cloudflare.py
